% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamrd.R
\name{bamrd}
\alias{bamrd}
\title{Reads the BAM file}
\usage{
bamrd(
  bam_file_name,
  reference_fasta,
  max_depth = 8000,
  min_base_quality = 13,
  min_mapq = 30,
  min_nucleotide_depth = 1,
  min_minor_allele_depth = 0,
  min_coverage = 0,
  use_noncanonical = TRUE,
  strand_specific = FALSE,
  single_end = TRUE
)
}
\arguments{
\item{bam_file_name}{A \code{character(1)} path to the input BAM file}

\item{reference_fasta}{A \code{character(1)} path to the reference genome FASTA file}

\item{max_depth}{An \code{integer(1)} maximum number of nucleotides to be included in pileup}

\item{min_base_quality}{An \code{integer(1)} minimum QUAL value for each nucleotide in an alignment}

\item{min_mapq}{An \code{integer(1)} minimum MAPQ value for an alignment to be included in pileup}

\item{min_nucleotide_depth}{An \code{integer(1)} minimum count of each nucleotide at a given position required for said nucleotide to appear in the result}

\item{min_minor_allele_depth}{An \code{integer(1)} left undocumented}

\item{min_coverage}{An \code{integer(1)} minimum coverage for position to be included in final dataset}

\item{use_noncanonical}{A \code{logical(1)} indicating  whether to use (\code{TRUE}) all chromosomes from BAM including non-canonical ones'}

\item{strand_specific}{A \code{logical(1)} indicating the presence either of the strand-specific (\code{TRUE}) RNA-seq experiment, or non-strand-specific (\code{FALSE}) one}

\item{single_end}{A \code{logical(1)} indicating the presence of the single-end (\code{TRUE}) RNA-seq experiment}
}
\value{
On \code{strand_specific == TRUE} the list of three data frames (with names \code{pileup_table}, \code{messy_positions}, and \code{pairs}), otherwise the list of only one data frame (with name \code{pileup_table}) suitable for the further RNA editing analysis
}
\description{
Separates data from the input BAM file into a number of data frames, each suitable for the further RNA editing analysis
}
\author{
Irina Shchyukina
}
