% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamrd.R
\name{fixed_chromosome_positions}
\alias{fixed_chromosome_positions}
\title{Actually get info from bam for given chromosome}
\usage{
fixed_chromosome_positions(
  bam_file,
  current_chromosome,
  bam_index,
  pileup_params,
  fasta_file,
  min_coverage,
  strand_specific,
  single_end
)
}
\description{
Actually get info from bam for given chromosome
}
\keyword{internal}
