% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamrd.R
\name{get_strand}
\alias{get_strand}
\title{Tries to define correct strand of the position via counting reads aligned to both strands}
\usage{
get_strand(
  pos_first,
  current_chromosome_length,
  current_chromosome,
  bam_file,
  single_end
)
}
\description{
Tries to define correct strand of the position via counting reads aligned to both strands
}
\keyword{internal}
