% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaedt.R
\name{rnaedt}
\alias{rnaedt}
\title{Performs RNA editing analysis}
\usage{
rnaedt(input, strand_specific = TRUE, threads = 20)
}
\arguments{
\item{input}{A list of the data frames returned by \code{bamrd} function}

\item{strand_specific}{A \code{logical(1)} indicating the presence either of the strand-specific (\code{TRUE}) RNA-seq experiment, or non-strand-specific (\code{FALSE}) one}

\item{threads}{An \code{integer(1)} number of threads}
}
\value{
Data frame with results of RNA editing analysis
}
\description{
Processes data supplied from \code{bamrd} and estimates RNA editing probability
}
\author{
Olga Dudkina
}
